Refactor test into two parts.
The following commit splits the tests into functionality testing
separate from authorization testing.  The former assumes that the
users have the permissions to do the action under test while the
latter checks that only the proper users get access to that action.
This should make it nice to keep the two types of tests separate from
each other.