Initial tests for aging advertisements.
These tests primarily test for being able to age advertisements.  They
require that the models support a new method (age for Board and Tile
and charge for Advertisement) so that the aging can be accomplished
properly.  The idea is that some other process (e.g., a cron job)
could run and call the age method on the proper models to cause them
to reset values and create a new charge for balance purposes.

More tests to follow...